Triple tang'd

+++ SuperSCAIANN/ANN.h
@@ -48,5 +48,5 @@ private:
 	const float eta = 0.01f; // regularization power
 
 	map< vector< vector<float> >, vector< vector<float> > > LoadedInputOutputData;
-	int argMaxIndex(vector<float> outputVector, int &index);
+	float argMaxIndex(vector<float> outputVector, int &index);
 };
